Move GraphQL documents to .gql file We do this for the Admin Generator, so we should do it here as well.
